Don't expose test functions
